test(ui): isolate suspension coverage
